requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.1.1
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.8
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
# pygraphviz==1.6
PyJWT==1.7.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
